% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaKTSP_Source.R
\name{meta.mul.rule.predict}
\alias{meta.mul.rule.predict}
\title{Prediction based on a given MetaKTSP model}
\usage{
meta.mul.rule.predict(test.dat, tspobj, K, is.youden = TRUE)
}
\arguments{
\item{test.dat}{Test dataset.}

\item{tspobj}{MetaKTSP model.}

\item{K}{The maximum number of top score pairs.}

\item{is.youden}{Used youden index in result. Default TRUE.}
}
\value{
An object of class list is returned with following three components:

"mul.rule" is the prediction result

"accuracy" is the prediction accuracy

"youden" is Youden index
}
\description{
Function for predicting subjects class based on given MetaKTSP model, and summarize the results
}
\author{
Chien-Wei Lin
}

