% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaKTSP_Source.R
\name{MetaTSP.mean}
\alias{MetaTSP.mean}
\title{A Meta-Analytic K Top Scoring Pair via Mean Score Method}
\usage{
MetaTSP.mean(DList, K.max, is.VO = TRUE, quantile.index = 10)
}
\arguments{
\item{DList}{Input variable matrix (a list of multiple datasets; row=features, column=samples).}

\item{K.max}{The maximum number of top score pairs.}

\item{is.VO}{Logical value indicating whether K selection is Variance optimization (VO)}

\item{quantile.index}{A quantile threshold for selecting top scoring pair. Default is 10.}
}
\value{
An object of class list is returned with components as following:
model includes meta k top scoring pairs (n by k-1).
num.K is the number of K meta top scoring pairs.
}
\description{
MetaTSP.mean is a function of a meta-analytic top scoring pair (MetaTSP) algorithm via Mean score method.
}
\author{
Chien-Wei Lin
}

