% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaKTSP_Source.R
\name{MetaTSP.Pvalue}
\alias{MetaTSP.Pvalue}
\title{A Meta-Analytic Top Scoring Pair Method for Robust Cross-Study Validation of Omics Prediction Analysis}
\usage{
MetaTSP.Pvalue(DList, Method = "Fisher", K.max, is.VO = TRUE,
  Para.num.gene = 10000)
}
\arguments{
\item{DList}{Input variable matrix (a list of multiple datasets; row=features, column=samples).}

\item{Method}{"Fisher" = fisher's meta-analysis, "Stouffer" = Stouffer's meta-analysis, NULL = mean score method.}

\item{K.max}{The maximum number of top score pairs.}

\item{is.VO}{Logical value indicating whether K selection is Variance optimization (VO).}

\item{Para.num.gene}{# of top scoring pairs that are initially considered.}
}
\value{
meta.tspobj = a list of meta top scoring pairs results
meta.tspobj$model = Meta top scoring pairs (n by k-1) meta.tspobj$num.K = #
of K meta top scoring pairs
}
\description{
Fit a prediction rule of meta top scoring pair
}
\author{
Chien-Wei Lin <masaki396@gmail.com>
}

