% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaKTSP_Source.R
\name{predict.mtsp}
\alias{predict.mtsp}
\title{Prediction based on a given MetaKTSP model}
\usage{
\method{predict}{mtsp}(object, test.dat, K)
}
\arguments{
\item{object}{MetaKTSP model.}

\item{test.dat}{A matrix with features in row and samples in column (p x n).}

\item{K}{The maximum number of top score pairs.}
}
\value{
A matrix of prediction result (K x n) with samples in column and different row correspond to different K.
}
\description{
Function for predicting subjects class based on given MetaKTSP model. For internal use in function meta.mul.rule.predict.
}
\author{
Chien-Wei Lin <masaki396@gmail.com>
}

